FIRE-710 Fix signup passport strategies
