+++ src/auth/strategy/github.strategy.ts
@@ -1,10 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-github2';
 import { CreateGithubUserDTO } from 'src/users/dto';
 import { AuthService } from '../auth.service';
 import { UsersDatabaseHelper } from '../../users/helper/users-database.helper';
+import { generateSignupVerifyToken } from 'src/users/helper/user.utils';
+import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class GithubStrategy extends PassportStrategy(Strategy, 'github') {
@@ -41,10 +43,12 @@ export class GithubStrategy extends PassportStrategy(Strategy, 'github') {
     if (!user) {
       try {
         // create user in modoco db
-        user = await this.usersDatabaseHelper.createGithubUser(
-          createUserDTO.nickname,
-          createUserDTO.email,
-          createUserDTO.githubId,
+        user = await this.createGithubUser(createUserDTO);
+        // send verification email
+        await this.emailService.sendVerificationMail(
+          user.uid,
+          user.email,
+          user.verify_token,
         );
       } catch (error) {
         done(error);
@@ -62,6 +66,29 @@ export class GithubStrategy extends PassportStrategy(Strategy, 'github') {
       ...user,
       access_token,
     });
+  }
+
+  private async createGithubUser(dto: CreateGithubUserDTO) {
+    try {
+      const verifyToken = generateSignupVerifyToken();
+
+      const user = await this.usersDatabaseHelper.createGithubUser(
+        dto.nickname,
+        dto.email,
+        verifyToken,
+        dto.githubId,
+      );
+
+      return user;
+    } catch (error) {
+      if (
+        error instanceof Prisma.PrismaClientKnownRequestError &&
+        error.code === 'P2002'
+      ) {
+        throw new ForbiddenException('User already exists');
+      }
+      throw error;
+    }
   }
 }
